Move authenticator to Cognite.Extensions
Makes it possible to use the utils for just authentication, not
requiring the whole config/metrics/state-store stack.
